increase heap and stack.  Fixed bad word detection.

+++ data_manager/DataManager.cpp
@@ -67,7 +67,7 @@ void DataManager::DumpDay() {
 	string outJson;
 	int count = 0;
 
-	for (int i = 0; i < size; i++)
+	for (int i = 0; i < (int)size; i++)
 	{
 		
 		stringstream ss;
